Consolidate CLIs in one place
