+++ Makefile
@@ -42,7 +42,7 @@ schema-diff: test-schema
 	rm prod.sql local.sql
 
 fake:
-	$(honcho_run) $(env_bin)/fake_data fake_data
+	$(honcho_run) $(env_bin)/fake-data
 
 run: env
 	PATH=$(env_bin):$(PATH) $(honcho_run) web
