+++ gratipay/cli/sync_npm.py
